+++ ui/src/main_window.hpp
@@ -28,6 +28,7 @@ class RasterViewerView;
 class ObjViewerView;
 class ProjectView;
 class ShpViewerView;
+class PbfViewerView;
 class TileGridView;
 
 class MainWindow : public QMainWindow {
@@ -42,7 +43,7 @@ protected:
     void dropEvent(QDropEvent* ev) override;
 
 private:
-    enum class FileKind { Unknown, Dsf, GeoTiff, Obj, Project, Shp };
+    enum class FileKind { Unknown, Dsf, GeoTiff, Obj, Project, Shp, Pbf };
 
     void build_menus();
     void build_toolbar();
@@ -69,6 +70,7 @@ private:
     ObjViewerView*    obj_view_       = nullptr;
     ProjectView*      project_view_   = nullptr;
     ShpViewerView*    shp_view_       = nullptr;
+    PbfViewerView*    pbf_view_       = nullptr;
     TileGridView*     map_view_       = nullptr;
 
     QPlainTextEdit*   log_edit_       = nullptr;